Add command service call to default talon controllers
Adds a "command" service call as an option for setting
the motor command for the default controller types
