+++ zebROS_ws/src/talon_controller_msgs/CMakeLists.txt
@@ -54,6 +54,7 @@ find_package(catkin REQUIRED COMPONENTS
 ## Generate services in the 'srv' folder
 add_service_files(
   FILES
+  Command.srv
   LoadMusicSrv.srv
   LoadInstrumentsSrv.srv
   PidfSlot.srv
